refactor: fsd barrel
