+++ src/refactoring/entities/coupon/index.ts
@@ -1,1 +1,2 @@
 export { DISCOUNT_TYPE, type DiscountType } from "./constants";
+export * from "./model";
