+++ src/refactoring/entities/product/model/index.ts
@@ -1,2 +1,3 @@
 export { createProduct, type Product } from "./product";
 export * from "./errors";
+export * from "./discount";
